split up librqbit torrent state into smaller files
