+++ crates/librqbit/src/lib.rs
@@ -6,7 +6,6 @@ pub mod http_api_client;
 mod http_api_error;
 pub mod peer_connection;
 pub mod peer_info_reader;
-pub mod peer_state;
 pub mod session;
 pub mod spawn_utils;
 pub mod torrent_manager;
