+++ crates/librqbit/src/torrent_state/mod.rs
@@ -1,3 +1,5 @@
+pub mod utils;
+
 pub mod live;
 
 pub use live::*;
